requirements.txt
requests
python-nmap
jinja2
PyYAML
python-whois
playwright  # optional, for screenshots
